changing nanonis-connection, not done

+++ src/connection/backing.rs
@@ -4,7 +4,7 @@ use nanonis_tcp::ScanDir;
 
 #[derive(Clone)]
 pub struct BufferState {
-    pub size: [usize; 2],
+    pub size: [u32; 2],
     pub buf_f: Arc<Vec<f32>>,
     pub buf_b: Arc<Vec<f32>>,
     pub scan_dir: ScanDir,
